feat: user 검증 로직 구현

+++ src/main/java/com/kt/service/UserServiceImpl.java
@@ -16,13 +16,10 @@ import com.kt.domain.dto.request.UserRequest;
 import com.kt.domain.dto.request.SignupRequest;
 import com.kt.domain.dto.response.OrderProductResponse;
 import com.kt.domain.dto.response.UserResponse;
-import com.kt.domain.entity.AbstractAccountEntity;
 import com.kt.domain.entity.OrderEntity;
 import com.kt.domain.entity.UserEntity;
-import com.kt.repository.account.AccountRepository;
 import com.kt.repository.orderproduct.OrderProductRepository;
 import com.kt.repository.OrderRepository;
-import com.kt.repository.review.ReviewRepository;
 import com.kt.repository.user.UserRepository;
 import com.kt.util.Preconditions;
 
@@ -37,13 +34,14 @@ public class UserServiceImpl implements UserService {
 	private final OrderRepository orderRepository;
 	private final UserRepository userRepository;
 	private final PasswordEncoder passwordEncoder;
-	private final AccountRepository accountRepository;
 
 
 	@Override
-	public UserResponse.Orders getOrdersByUserId(UUID id) {
-		List<OrderEntity> orders = orderRepository.findAllByOrderBy_Id(id);
-		return UserResponse.Orders.of(id, orders);
+	public UserResponse.Orders getOrdersByUserId(UUID currentId, UUID subjectId) {
+		verifyAccess(currentId, subjectId);
+
+		List<OrderEntity> orders = orderRepository.findAllByOrderBy_Id(subjectId);
+		return UserResponse.Orders.of(subjectId, orders);
 	}
 
 	@Override
@@ -52,13 +50,16 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public Page<UserResponse.Search> getUsers(Pageable pageable, String keyword, UserRole role) {
+	public Page<UserResponse.Search> getUsers(UUID userId, Pageable pageable, String keyword, UserRole role) {
+		verifyAdmin(userId);
 		return userRepository.searchUsers(pageable, keyword, role);
 	}
 
 	@Override
-	public UserResponse.UserDetail getUserDetail(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public UserResponse.UserDetail getUserDetail(UUID currentId, UUID subjectId) {
+		verifyAccess(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		return new UserResponse.UserDetail(
 			user.getId(),
 			user.getName(),
@@ -71,8 +72,8 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public UserResponse.UserDetail getAdminDetail(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public UserResponse.UserDetail getUserDetailSelf(UUID currentId) {
+		UserEntity user = userRepository.findByIdOrThrow(currentId);
 		return new UserResponse.UserDetail(
 			user.getId(),
 			user.getName(),
@@ -85,31 +86,57 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public void enableUser(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public UserResponse.UserDetail getAdminDetail(UUID currentId, UUID subjectId) {
+		verifyAccess(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
+		return new UserResponse.UserDetail(
+			user.getId(),
+			user.getName(),
+			user.getEmail(),
+			user.getRole(),
+			user.getGender(),
+			user.getBirth(),
+			user.getMobile()
+		);
+	}
+
+	@Override
+	public void enableUser(UUID currentId, UUID subjectId) {
+		verifyAccess(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		user.enabled();
 	}
 
 	@Override
-	public void disableUser(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public void disableUser(UUID currentId, UUID subjectId) {
+		verifyAccess(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		user.disabled();
 	}
 
 	@Override
-	public void deleteUser(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public void deleteUser(UUID currentId, UUID subjectId) {
+		verifyAccess(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		user.delete();
 	}
 
 	@Override
-	public void retireUser(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public void retireUser(UUID currentId, UUID subjectId) {
+		verifyAccess(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		user.retired();
 	}
 
 	@Override
-	public void createAdmin(SignupRequest.SignupMember request) {
+	public void createAdmin(UUID userId, SignupRequest.SignupMember request) {
+		verifyAdmin(userId);
+
 		UserEntity admin = UserEntity.create(
 			request.name(),
 			request.email(),
@@ -123,27 +150,33 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public void deleteUserPermanently(UUID id) {
-		UserEntity user = userRepository.findByIdOrThrow(id);
+	public void deleteUserPermanently(UUID currentId, UUID subjectId) {
+		verifyAccess(currentId, subjectId);
+
+		UserEntity user = userRepository.findByIdOrThrow(subjectId);
 		orderRepository.clearUser(user.getId());
 		userRepository.delete(user);
 	}
 
 	@Override
-	public void deleteAdmin(UUID adminId) {
+	public void deleteAdmin(UUID currentId, UUID adminId) {
+		verifyAccess(currentId, adminId);
+
 		UserEntity user = userRepository.findByIdOrThrow(adminId);
 		user.delete();
 	}
 
 	@Override
 	public void updateUserDetail(
-		UUID currentUserId,
-		UUID userId,
+		UUID currentId,
+		UUID subjectId,
 		UserRequest.UpdateDetails details
 	) {
-		UserEntity user = userRepository.findByIdOrThrow(userId);
-		hasUserAccessPermission(currentUserId, user);
-		user.updateDetails(
+		verifyAccess(currentId, subjectId);
+
+		UserEntity subjectUser = userRepository.findByIdOrThrow(subjectId);
+
+		subjectUser.updateDetails(
 			details.name(),
 			details.mobile(),
 			details.birth(),
@@ -151,11 +184,43 @@ public class UserServiceImpl implements UserService {
 		);
 	}
 
-	private void hasUserAccessPermission(UUID currentUserId, UserEntity user){
-		AbstractAccountEntity userEditor = accountRepository.findByIdOrThrow(currentUserId);
+	@Override
+	public void updateUserDetailSelf(
+		UUID userId,
+		UserRequest.UpdateDetails details
+	) {
+		UserEntity subjectUser = userRepository.findByIdOrThrow(userId);
+
+		subjectUser.updateDetails(
+			details.name(),
+			details.mobile(),
+			details.birth(),
+			details.gender()
+		);
+	}
+
+	private void verifyAdmin(UUID currentUserId) {
+		UserEntity user = userRepository.findByIdOrThrow(currentUserId);
 		Preconditions.validate(
-			userEditor.getEmail().equals(user.getEmail()),
+			user.getRole() == UserRole.ADMIN,
 			ErrorCode.ACCOUNT_ACCESS_NOT_ALLOWED
 		);
+	}
+
+	private void verifyAccess(UUID currentId, UUID subjectId) {
+		UserEntity subjectUser = userRepository.findByIdOrThrow(subjectId);
+
+		if (subjectUser.getRole() == UserRole.ADMIN) {
+			Preconditions.validate(
+				currentId.equals(subjectId),
+				ErrorCode.ACCOUNT_ACCESS_NOT_ALLOWED
+			);
+		} else {
+			UserEntity currentUser = userRepository.findByIdOrThrow(currentId);
+			Preconditions.validate(
+				currentUser.getRole().equals(UserRole.ADMIN) || currentId.equals(subjectId),
+				ErrorCode.ACCOUNT_ACCESS_NOT_ALLOWED
+			);
+		}
 	}
 }
